MP-58 :recycle: Refactor: RedisTemplate 설정 리팩토링 및 직렬화 로직 개선
- StringRedisSerializer 중복 생성 제거 및 변수로 추출
- ObjectMapper 설정 방식 configure → disable로 통일
- GenericJackson2JsonRedisSerializer 설정을 메소드 분리로 정리
